Conflicts:
	PerfMetrics.py

+++ Utils.py
@@ -24,7 +24,7 @@ def read_train_trps_txt(path, toNX=False, skip = 0, fn = ''):
         fname = TRAIN_FNAME
     if toNX:
         return nx.read_weighted_edgelist(path + fname, create_using=nx.DiGraph(), nodetype=int)
-    return np.loadtxt(path + fname, dtype='int32', skiprows = skip)
+    return np.loadtxt(path + fname, skiprows = skip)
 
 def read_test_trps_txt(path, toNX=False, skip = 0):
 # Accepts path to TEST_FNAME
@@ -32,7 +32,7 @@ def read_test_trps_txt(path, toNX=False, skip = 0):
 # Can be given a number of rows to skip, ndarray case only
     if toNX:
         return nx.read_weighted_edgelist(path + TEST_FNAME, create_using=nx.DiGraph(), nodetype=int)
-    return np.loadtxt(path + TEST_FNAME, dtype='int32', skiprows = skip)
+    return np.loadtxt(path + TEST_FNAME, skiprows = skip)
 
 def get_coo(trp_raw):
 # Given nX3 raw trp ndarray trp_raw, return a sparse.coo_matrix of minimal square shape
